+++ tests/clock.test.js
@@ -42,7 +42,7 @@ function assertNotNull(value, message) {
 
 // Load HTML and create DOM
 function createDOM() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     resources: 'usable',
@@ -496,7 +496,7 @@ console.log('----------------------------');
   const dom = createDOM();
   const { document } = dom.window;
 
-  assertEqual(document.title, 'TaskTimer', 'Default tab title is "TaskTimer"');
+  assertEqual(document.title, 'TimerPie', 'Default tab title is "TimerPie"');
 
   dom.window.close();
 })();
@@ -512,9 +512,9 @@ console.log('----------------------------');
   timeInput.dispatchEvent(new dom.window.Event('input'));
   goBtn.click();
 
-  // Title should show countdown (format: "M:SS - TaskTimer" or similar)
+  // Title should show countdown (format: "M:SS - TimerPie" or similar)
   assert(document.title.includes(':'), 'Tab title shows time format while running');
-  assert(document.title.includes('TaskTimer'), 'Tab title still includes TaskTimer');
+  assert(document.title.includes('TimerPie'), 'Tab title still includes TimerPie');
 
   dom.window.close();
 })();
@@ -1209,7 +1209,7 @@ console.log('-------------------------------------');
 
 (function testPreferencesRestoredFromCookies() {
   // Create DOM with pre-set cookies
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const cookieJar = new (require('jsdom').CookieJar)();
   cookieJar.setCookieSync('clockColor=#4a90e2', 'http://localhost');
   cookieJar.setCookieSync('clockMode=cw', 'http://localhost');
@@ -1300,7 +1300,7 @@ console.log('-------------------------------------------');
 
 (function testCookieValueValidation() {
   // Test that invalid color values are rejected when restoring
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const cookieJar = new (require('jsdom').CookieJar)();
   // Set invalid color value
   cookieJar.setCookieSync('clockColor=invalid', 'http://localhost');
@@ -1328,7 +1328,7 @@ console.log('Integration Tests: Tooltips');
 console.log('----------------------------');
 
 (function testStaticTooltips() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1358,7 +1358,7 @@ console.log('----------------------------');
 })();
 
 (function testDynamicTooltips() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1396,7 +1396,7 @@ console.log('Integration Tests: Keyboard Shortcuts');
 console.log('--------------------------------------');
 
 (function testKeyboardShortcuts() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1429,7 +1429,7 @@ console.log('--------------------------------------');
 })();
 
 (function testSpaceIgnoredInEndMode() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1461,7 +1461,7 @@ console.log('Integration Tests: URL Sharing');
 console.log('-------------------------------');
 
 (function testUrlHashUpdate() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1487,7 +1487,7 @@ console.log('-------------------------------');
 })();
 
 (function testUrlHashParsing() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1530,7 +1530,7 @@ console.log('--------------------------------------');
 })();
 
 (function testPipButtonExists() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1555,7 +1555,7 @@ console.log('Integration Tests: Alarm Sound');
 console.log('-------------------------------');
 
 (function testSoundButtonExists() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1569,7 +1569,7 @@ console.log('-------------------------------');
 })();
 
 (function testSoundToggle() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document, ClockLogic } = dom.window;
 
@@ -1602,7 +1602,7 @@ console.log('-------------------------------');
 })();
 
 (function testPlayAlarmFunctionExists() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { ClockLogic } = dom.window;
 
@@ -1617,7 +1617,7 @@ console.log('Integration Tests: PWA Manifest');
 console.log('--------------------------------');
 
 (function testManifestLinkExists() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1644,7 +1644,7 @@ console.log('--------------------------------');
 })();
 
 (function testPWAMetaTags() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1783,7 +1783,7 @@ console.log('Security: URL Hash Input Validation');
 console.log('------------------------------------');
 
 (function testTimeHashParameterAcceptsValidInput() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1797,7 +1797,7 @@ console.log('------------------------------------');
 })();
 
 (function testTimeHashParameterAcceptsColonFormat() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1811,7 +1811,7 @@ console.log('------------------------------------');
 })();
 
 (function testTimeHashParameterAcceptsDecimal() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1825,7 +1825,7 @@ console.log('------------------------------------');
 })();
 
 (function testGoButtonEnabledWithTimeFromURL() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, {
     runScripts: 'dangerously',
     pretendToBeVisual: true,
@@ -1893,7 +1893,7 @@ console.log('Integration Tests: Wake Lock');
 console.log('-----------------------------');
 
 (function testWakeLockFunctionsExist() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { ClockLogic } = dom.window;
 
@@ -1904,7 +1904,7 @@ console.log('-----------------------------');
 })();
 
 (function testWakeLockReleaseClearsReference() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { ClockLogic } = dom.window;
 
@@ -1919,7 +1919,7 @@ console.log('-----------------------------');
 })();
 
 (function testVisibilityChangeHandlerExists() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
 
   // Check that visibilitychange event listener is registered
   assert(html.includes('visibilitychange'), 'Visibility change listener registered');
@@ -1934,7 +1934,7 @@ console.log('Accessibility Tests');
 console.log('-------------------');
 
 (function testSVGAccessibility() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1949,7 +1949,7 @@ console.log('-------------------');
 })();
 
 (function testARIALiveRegion() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;
 
@@ -1962,7 +1962,7 @@ console.log('-------------------');
 })();
 
 (function testPresetButtonAriaLabels() {
-  const html = fs.readFileSync(path.join(__dirname, '..', 'TaskTimer.html'), 'utf8');
+  const html = fs.readFileSync(path.join(__dirname, '..', 'TimerPie.html'), 'utf8');
   const dom = new JSDOM(html, { runScripts: 'dangerously', pretendToBeVisual: true });
   const { document } = dom.window;